Bump version to try to register latest code
The last commit on the master branch is not what the general registry tells
Pkg to install. Instead it installs the previous commit. I don't know why.
I can't find the information in the general registry that says which commit to
use. This is mysterious. I will try to bump the version number and register
the new version and see if the latest code is then installed via Pkg.